Remove MUC lines instead of restoring file backup
This way, the user is free to make changes without fear of losing what they did when they manually run MUC -cleanup (even though that is not a common use case).

Backups of the scripts which we touch will now be stored in the MUC_Backup folder. Modified and unmodified backups will be kept individually for both files. 99% of what could go wrong is only possible if the user is running the exe manually instead of via a script which runs it twice in quick succession.